Add Select API docs to storybook
